chore(lib.rs): Separated Config type into a separate file
