+++ src/lib.rs
@@ -1,75 +1,10 @@
-pub mod config {
-    use serde::Deserialize;
-
-    use std::error::Error;
-    use std::fs::File;
-    use std::io::BufReader;
-    use std::path::Path;
-
-    #[derive(Deserialize, Debug)]
-    pub struct Config {
-        pub db_address: String,
-    }
-
-    pub fn read_json_config(path: impl AsRef<Path>) -> Result<Config, Box<dyn Error>> {
-        let file = File::open(path)?;
-        let reader = BufReader::new(file);
-
-        // read json from file
-        let json_data = serde_json::from_reader(reader)?;
-
-        Ok(json_data)
-    }
-}
-
-pub mod db {
-    use chrono::{
-        serde::{ts_seconds, ts_seconds_option},
-        DateTime, Utc,
-    };
-    use mongodb::{bson::doc, sync::Collection};
-    use serde::{Deserialize, Serialize};
-
-    #[derive(Debug, Serialize, Deserialize)]
-    // Task型
-    pub struct Task {
-        #[serde(with = "ts_seconds_option")]
-        pub scheduled_date: Option<DateTime<Utc>>,
-        #[serde(with = "ts_seconds")]
-        pub register_date: DateTime<Utc>,
-        pub title: String,
-        pub memo: String,
-        pub finish: bool,
-    }
-
-    impl Default for Task {
-        fn default() -> Self {
-            Self {
-                scheduled_date: None,
-                register_date: Utc::now(),
-                title: String::new(),
-                memo: String::new(),
-                finish: false,
-            }
-        }
-    }
-
-    //新規タスクの追加
-    pub fn add_task(
-        new_task: Task,
-        collection: &Collection<Task>,
-    ) -> Result<(), mongodb::error::Error> {
-        collection.insert_one(new_task, None)?;
-        Ok(())
-    }
-}
-
+pub mod config;
+pub mod mongodb;
+pub mod monoglossic_repository;
 #[cfg(test)]
+#[cfg(FALSE)]
 mod tests {
-    use crate::{
-        config::{read_json_config, Config},
-        db::{add_task, Task},
-    };
+    use crate::config::{read_json_config, Config};
     use chrono::{TimeZone, Utc};
     use mongodb::sync::Client;
